+++ doc/changelog.rst
@@ -22,6 +22,7 @@ Localization
 Admin
 ~~~~~
 
+- Add picotable select2 and MPTT filters
 - Hide cancelled orders by default from orders lists
 - Add option to delete shipments
 - Apply picotable text filters on change rather than on enter/on focus out